Add Partition_statistics_type for comprehensive MBR partition statistics
